ai-service/requirements.txt
albucore==0.0.24
albumentations==2.0.8
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
boto3==1.42.47
botocore==1.42.47
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
coloredlogs==15.0.1
contourpy==1.3.0
cycler==0.12.1
Cython==3.2.4
easydict==1.13
eval_type_backport==0.3.1
exceptiongroup==1.3.1
faiss-cpu==1.13.0
fastapi==0.128.8
flatbuffers==25.12.19
fonttools==4.60.2
h11==0.16.0
humanfriendly==10.0
idna==3.11
ImageIO==2.37.2
importlib_resources==6.5.2
insightface==0.7.3
jmespath==1.1.0
joblib==1.5.3
kiwisolver==1.4.7
lazy_loader==0.4
matplotlib==3.9.4
ml_dtypes==0.5.4
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
onnx==1.19.1
onnxruntime==1.19.2
opencv-python-headless==4.13.0.92
packaging==26.0
pillow==11.3.0
prettytable==3.16.0
protobuf==6.33.5
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
redis==7.0.1
requests==2.32.5
s3transfer==0.16.0
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
simsimd==6.5.12
six==1.17.0
starlette==0.49.3
stringzilla==4.6.0
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2024.8.30
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==1.26.20
uvicorn==0.39.0
wcwidth==0.6.0
zipp==3.23.0
